Move font sources to top level and hint at UFO format
This is a git repository and (IMO) the primary reason for existing
should be to keep the font sources. Everything else is secondary, hence
I don't see the utility of stuffing *the main thing* in a subdirectory.

The .ufo naming is a nod to non-Mac platforms that treat UFO sources as
a directory. The hint that the directory *is* the UFO makes it easier on
users that could easily get confused, especially when there is
a "glyphs" subdirectory!